src/main.rs: add --each-line, and --include-input flags
--each-line flag
Suitable for process the input from file in which it will process
each line from the file. Thus each line is independent to each
other. It is usually suitable for batch processing.

--include-input
This will include the input text for the output as well.
Output is in the format of the following

<input_text> <output_text>

wheres input_text follows by a space, then follows by output_text.